ISSUE #59 * Rename `tests.modelClasses` module as `tests.model_classes`.
